pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scholar-watch"
version = "0.1.0"
description = "Track citation performance of researchers on Google Scholar"
requires-python = ">=3.10"
dependencies = [
    "scholarly>=1.7",
    "SQLAlchemy>=2.0",
    "Flask>=3.0",
    "flask-login>=0.6",
    "plotly>=5.18",
    "kaleido>=0.2",
    "PyYAML>=6.0",
    "Jinja2>=3.1",
    "bcrypt>=4.0",
    "gunicorn>=22.0",
    "python-dotenv>=1.0",
    "alembic>=1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
scholar-watch = "scholar_watch.cli:main"

[tool.setuptools.packages.find]
include = ["scholar_watch*"]
